refactor: Complete Phase 2b - Extract all remaining handlers to domain modules
Phase 2b extracts ALL remaining complex handlers from fantasy_football_multi_league.py
into dedicated, domain-organized modules. This completes the handler extraction
initiative started in Phase 2a.

Changes:
- Main file reduced from 2,154 → 1,155 lines (46% reduction, 999 lines extracted)
- Created 5 new handler modules organized by domain
- Implemented comprehensive dependency injection pattern
- Maintained 100% test pass rate (70/70 tests passing)

New Handler Modules:
1. roster_handlers.py (212 lines)
   - handle_ff_get_roster: Complex roster retrieval with enhanced data

2. matchup_handlers.py (205 lines)
   - handle_ff_get_matchup: Matchup information
   - handle_ff_build_lineup: Advanced lineup optimization
   - handle_ff_compare_teams: Team comparison

3. player_handlers.py (572 lines)
   - handle_ff_get_players: Top available players
   - handle_ff_get_waiver_wire: Comprehensive waiver analysis (largest handler)

4. draft_handlers.py (128 lines)
   - handle_ff_get_draft_results: Draft results
   - handle_ff_get_draft_rankings: Draft rankings with ADP
   - handle_ff_get_draft_recommendation: Live draft recommendations
   - handle_ff_analyze_draft_state: Draft state analysis

5. analytics_handlers.py (23 lines)
   - handle_ff_analyze_reddit_sentiment: Reddit sentiment analysis

Dependency Injection:
- Updated src/handlers/__init__.py with injection functions for each domain
- inject_roster_dependencies(), inject_matchup_dependencies(), etc.
- Moved injection timing to after helper functions are defined (line 1105+)
- All dependencies explicitly documented in handler modules

Architecture:
- Main file now contains ONLY orchestration layer
- Server setup, helper functions, tool definitions, dependency injection
- All handler logic cleanly separated by domain
- Clear boundaries enable independent testing and development

Testing:
- All 70 unit + integration tests passing
- No regressions from refactoring
- Tests validate dependency injection works correctly

Documentation:
- Created PHASE_2B_REFACTOR_SUMMARY.md with detailed analysis
- Updated CLAUDE.md with new handler organization
- Documented dependency injection pattern and architecture

Total Refactoring Progress:
- Started: 2,675 lines (monolith)
- After Phase 1 + 2a: 2,154 lines
- After Phase 2b: 1,155 lines (57% total reduction)
