+++ src/handlers/__init__.py
@@ -1,13 +1,13 @@
-"""MCP tool handlers - imports from handler modules and main file."""
+"""MCP tool handlers - orchestrates handler modules with dependency injection."""
 
-# Import simple handlers from dedicated modules
+# Import simple handlers from dedicated modules (no dependencies)
 from .admin_handlers import (
     handle_ff_clear_cache,
     handle_ff_get_api_status,
     handle_ff_refresh_token,
 )
 
-# League handlers will need helper functions injected
+# League handlers (need helper function injection)
 from .league_handlers import (
     handle_ff_get_league_info,
     handle_ff_get_leagues,
@@ -15,36 +15,90 @@ from .league_handlers import (
     handle_ff_get_teams,
 )
 
-# Complex handlers - will be imported from main file initially
-# Then gradually refactored in future sessions
-_handle_ff_get_roster = None
-_handle_ff_get_matchup = None
-_handle_ff_build_lineup = None
-_handle_ff_get_players = None
-_handle_ff_compare_teams = None
-_handle_ff_get_waiver_wire = None
-_handle_ff_get_draft_results = None
-_handle_ff_get_draft_rankings = None
-_handle_ff_get_draft_recommendation = None
-_handle_ff_analyze_draft_state = None
-_handle_ff_analyze_reddit_sentiment = None
+# Roster handlers (need dependency injection)
+from .roster_handlers import handle_ff_get_roster
+
+# Matchup handlers (need dependency injection)
+from .matchup_handlers import (
+    handle_ff_build_lineup,
+    handle_ff_compare_teams,
+    handle_ff_get_matchup,
+)
+
+# Player handlers (need dependency injection)
+from .player_handlers import (
+    handle_ff_get_players,
+    handle_ff_get_waiver_wire,
+)
+
+# Draft handlers (need dependency injection)
+from .draft_handlers import (
+    handle_ff_analyze_draft_state,
+    handle_ff_get_draft_rankings,
+    handle_ff_get_draft_recommendation,
+    handle_ff_get_draft_results,
+)
+
+# Analytics handlers (minimal dependencies)
+from .analytics_handlers import handle_ff_analyze_reddit_sentiment
 
 
-def inject_complex_handlers(**handlers):
-    """Inject complex handler functions from main module.
+def inject_roster_dependencies(**deps):
+    """Inject dependencies needed by roster handlers.
 
-    This allows us to gradually refactor complex handlers while keeping
-    the system working. As handlers are refactored, they can be moved
-    to dedicated modules.
+    Required dependencies:
+    - get_user_team_info: Get user's team info in a league
+    - yahoo_api_call: Make Yahoo API calls
+    - parse_team_roster: Parse roster from Yahoo API response
     """
-    global _handle_ff_get_roster, _handle_ff_get_matchup, _handle_ff_build_lineup
-    global _handle_ff_get_players, _handle_ff_compare_teams, _handle_ff_get_waiver_wire
-    global _handle_ff_get_draft_results, _handle_ff_get_draft_rankings
-    global _handle_ff_get_draft_recommendation, _handle_ff_analyze_draft_state
-    global _handle_ff_analyze_reddit_sentiment
+    import src.handlers.roster_handlers as roster_mod
 
-    for name, func in handlers.items():
-        globals()[name] = func
+    for name, func in deps.items():
+        setattr(roster_mod, name, func)
+
+
+def inject_matchup_dependencies(**deps):
+    """Inject dependencies needed by matchup handlers.
+
+    Required dependencies:
+    - get_user_team_key: Get user's team key in a league
+    - get_user_team_info: Get user's team info in a league
+    - yahoo_api_call: Make Yahoo API calls
+    - parse_team_roster: Parse roster from Yahoo API response
+    """
+    import src.handlers.matchup_handlers as matchup_mod
+
+    for name, func in deps.items():
+        setattr(matchup_mod, name, func)
+
+
+def inject_player_dependencies(**deps):
+    """Inject dependencies needed by player handlers.
+
+    Required dependencies:
+    - yahoo_api_call: Make Yahoo API calls
+    - get_waiver_wire_players: Get waiver wire players
+    """
+    import src.handlers.player_handlers as player_mod
+
+    for name, func in deps.items():
+        setattr(player_mod, name, func)
+
+
+def inject_draft_dependencies(**deps):
+    """Inject dependencies needed by draft handlers.
+
+    Required dependencies:
+    - get_all_teams_info: Get all teams info
+    - get_draft_rankings: Get draft rankings
+    - get_draft_recommendation_simple: Get draft recommendations
+    - analyze_draft_state_simple: Analyze draft state
+    - DRAFT_AVAILABLE: Draft availability flag
+    """
+    import src.handlers.draft_handlers as draft_mod
+
+    for name, value in deps.items():
+        setattr(draft_mod, name, value)
 
 
 def inject_league_helpers(**helpers):
@@ -64,21 +118,31 @@ __all__ = [
     "handle_ff_refresh_token",
     "handle_ff_get_api_status",
     "handle_ff_clear_cache",
-    # League handlers (extracted but need helper injection)
+    # League handlers (extracted, need helper injection)
     "handle_ff_get_leagues",
     "handle_ff_get_league_info",
     "handle_ff_get_standings",
     "handle_ff_get_teams",
-    # Complex handlers (to be extracted in future)
-    "_handle_ff_get_roster",
-    "_handle_ff_get_matchup",
-    "_handle_ff_build_lineup",
-    "_handle_ff_get_players",
-    "_handle_ff_compare_teams",
-    "_handle_ff_get_waiver_wire",
-    "_handle_ff_get_draft_results",
-    "_handle_ff_get_draft_rankings",
-    "_handle_ff_get_draft_recommendation",
-    "_handle_ff_analyze_draft_state",
-    "_handle_ff_analyze_reddit_sentiment",
+    # Roster handlers (extracted, need dependency injection)
+    "handle_ff_get_roster",
+    # Matchup handlers (extracted, need dependency injection)
+    "handle_ff_get_matchup",
+    "handle_ff_build_lineup",
+    "handle_ff_compare_teams",
+    # Player handlers (extracted, need dependency injection)
+    "handle_ff_get_players",
+    "handle_ff_get_waiver_wire",
+    # Draft handlers (extracted, need dependency injection)
+    "handle_ff_get_draft_results",
+    "handle_ff_get_draft_rankings",
+    "handle_ff_get_draft_recommendation",
+    "handle_ff_analyze_draft_state",
+    # Analytics handlers (extracted, minimal dependencies)
+    "handle_ff_analyze_reddit_sentiment",
+    # Injection functions
+    "inject_roster_dependencies",
+    "inject_matchup_dependencies",
+    "inject_player_dependencies",
+    "inject_draft_dependencies",
+    "inject_league_helpers",
 ]
